refactor: employee api folder structure
