+++ src/features/employees/api/index.ts
@@ -1,8 +1,7 @@
-export { default as usePatchEmployee } from "./usePatchEmployee";
-export { default as useAssociateEmployeeWithInstance } from "./useAssociateEmployeeWithInstance";
-export { default as useDisassociateEmployeeFromInstance } from "./useDisassociateEmployeeFromInstance";
-export { default as useGetEmployees } from "./useGetEmployees";
-export { default as useGetEmployeesInfinite } from "./useGetEmployeesInfinite";
-export { default as useGetEmployee } from "./useGetEmployee";
-export { default as useDeleteEmployee } from "./useDeleteEmployee";
-export { default as useDeactivateEmployee } from "./useDeactivateEmployee";
+export { usePatchEmployee } from "./usePatchEmployee";
+export { useAssociateEmployeeWithInstance } from "./useAssociateEmployeeWithInstance";
+export { useDisassociateEmployeeFromInstance } from "./useDisassociateEmployeeFromInstance";
+export { useGetEmployees } from "./useGetEmployees";
+export { useGetEmployeesInfinite } from "./useGetEmployeesInfinite";
+export { useGetEmployee } from "./useGetEmployee";
+export { useDeactivateEmployee } from "./useDeactivateEmployee";
