--- src/features/employees/api/useAssociateEmployeeWithInstance/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./useAssociateEmployeeWithInstance";
