+++ src/features/employees/api/useDeactivateEmployee.ts
@@ -2,15 +2,15 @@ import useFetch from "@/hooks/useFetch";
 import type { ApiError } from "@/types/api/ApiError";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import type { AxiosError, AxiosResponse } from "axios";
-import type { Employee } from "../../types";
-import type { EmployeeActionParams } from "../../types/EmployeeParams";
+import type { Employee } from "../types";
+import type { EmployeeActionParams } from "../types/EmployeeParams";
 
 interface DeactivateEmployeeParams extends EmployeeActionParams {
   sanitize_instances?: boolean;
   remove_instances?: boolean;
 }
 
-const useDeactivateEmployee = () => {
+export const useDeactivateEmployee = () => {
   const authFetch = useFetch();
   const queryClient = useQueryClient();
 
@@ -32,5 +32,3 @@ const useDeactivateEmployee = () => {
     isDeactivating: isPending,
   };
 };
-
-export default useDeactivateEmployee;
