--- src/features/employees/api/useDeactivateEmployee/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./useDeactivateEmployee";
