--- src/features/employees/api/useDeleteEmployee/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./useDeleteEmployee";
