+++ src/features/employees/api/useDisassociateEmployeeFromInstance.ts
@@ -2,9 +2,9 @@ import useFetch from "@/hooks/useFetch";
 import type { ApiError } from "@/types/api/ApiError";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import type { AxiosError, AxiosResponse } from "axios";
-import type { EmployeeActionWithInstanceParams } from "../../types";
+import type { EmployeeActionWithInstanceParams } from "../types";
 
-const useDisassociateEmployeeFromInstance = () => {
+export const useDisassociateEmployeeFromInstance = () => {
   const authFetch = useFetch();
   const queryClient = useQueryClient();
 
@@ -28,5 +28,3 @@ const useDisassociateEmployeeFromInstance = () => {
     isDisassociating: isPending,
   };
 };
-
-export default useDisassociateEmployeeFromInstance;
