--- src/features/employees/api/useDisassociateEmployeeFromInstance/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./useDisassociateEmployeeFromInstance";
