--- src/features/employees/api/useGetEmployee/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./useGetEmployee";
