--- src/features/employees/api/useGetEmployees/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./useGetEmployees";
