--- src/features/employees/api/useGetEmployeesInfinite/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./useGetEmployeesInfinite";
