--- src/features/employees/api/usePatchEmployee/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./usePatchEmployee";
